feat: create segment to display Terraform version
This displays the current version of Terraform
when Terraform files are detected in current directory.

The directories to be searched are configurable
as well as the file extensions used for detection

To improve execution times, it reads version pinning configs from 'tfenv'
and 'asdf' if present before getting the version string
from the Terraform binary directly.